+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 v0.1.3 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Updated WithRangeConsumers optional to specify TopicType.
 
 
 v0.1.2 (2018-03-07)